Fix predicting
viterbi.run_viterbi() would always load the files to run viterbi.py on from a hard-coded path. This hard coded path pointed to the CNN output from running testing.py. Thus, prediction results would not be to the CNN output for data in DATA/PREDICTING, but for data in DATA/TESTING. The path to the raw CNN output is now a parameter of the function instead.